refactor(tii): extract type->JSON Schema mapping into schema module
Move SchemaCtx, the env/param schema inference, and their tests out of
tii/mod.rs into a dedicated tii/schema.rs. The mapping depends only on
the AST and serde_json, so it decouples cleanly and sits alongside the
existing tii/types.rs.
